+++ tests/scriptlets/m3u-prune.test.js
@@ -369,7 +369,6 @@ if (!isSupported) {
                 xhr.responseText.includes('tvessaiprod.nbcuni.com/video/'),
                 'line with "tvessaiprod.nbcuni.com/video/" should not be removed',
             );
-            assert.strictEqual(window.hit, undefined, 'should not hit');
             done();
         };
         xhr.send();
@@ -386,7 +385,6 @@ if (!isSupported) {
         xhr.open(METHOD, M3U8_PATH);
         xhr.onload = () => {
             assert.ok(xhr.responseText.includes('#EXT-X-VMAP-AD-BREAK'));
-            assert.strictEqual(window.hit, undefined, 'should not hit');
             done();
         };
         xhr.send();
@@ -410,7 +408,6 @@ if (!isSupported) {
                 xhr.responseText.includes('tvessaiprod.nbcuni.com/video/'),
                 'line with "tvessaiprod.nbcuni.com/video/" should not be removed',
             );
-            assert.strictEqual(window.hit, undefined, 'should not hit');
             done();
         };
         xhr.send();
@@ -434,7 +431,6 @@ if (!isSupported) {
                 xhr.responseText.includes('#EXT-X-VMAP-AD-BREAK'),
                 'line with "#EXT-X-VMAP-AD-BREAK" should not be removed',
             );
-            assert.strictEqual(window.hit, undefined, 'should not hit');
             done();
         };
         xhr.send();
@@ -659,5 +655,74 @@ if (!isSupported) {
             done();
         };
         xhr.send();
+    });
+
+    test('xhr - remove ads - addEventListener', async (assert) => {
+        const METHOD = 'GET';
+        const M3U8_PATH = M3U8_OBJECTS_PATH_01;
+        const MATCH_DATA = 'tvessaiprod.nbcuni.com/video/';
+
+        runScriptlet(name, [MATCH_DATA]);
+
+        const done = assert.async();
+
+        const xhr = new XMLHttpRequest();
+        xhr.open(METHOD, M3U8_PATH);
+        xhr.addEventListener('readystatechange', () => {
+            if (xhr.responseText && xhr.readyState >= 3) {
+                assert.notOk(
+                    xhr.responseText.includes('tvessaiprod.nbcuni.com/video/'),
+                    'check if "tvessaiprod.nbcuni.com/video/" has been removed',
+                );
+                assert.notOk(
+                    xhr.responseText.includes('#EXT-X-CUE:TYPE="SpliceOut"'),
+                    'check if "#EXT-X-CUE:TYPE="SpliceOut"" has been removed',
+                );
+                assert.notOk(
+                    xhr.responseText.includes('#EXT-X-CUE-IN'),
+                    'check if "#EXT-X-CUE-IN" has been removed',
+                );
+                assert.notOk(
+                    xhr.responseText.includes('#EXT-X-ASSET:CAID'),
+                    'check if "#EXT-X-ASSET:CAID" has been removed',
+                );
+                assert.notOk(
+                    xhr.responseText.includes('#EXT-X-SCTE35:'),
+                    'check if "#EXT-X-SCTE35:" has been removed',
+                );
+            }
+        });
+        xhr.onload = () => {
+            assert.strictEqual(window.hit, 'FIRED', 'hit function fired');
+            done();
+        };
+        xhr.send();
+    });
+
+    // This test is for issue with - Request with body": Failed to execute 'fetch' on 'Window':
+    // Cannot construct a Request with a Request object that has already been used
+    test('fetch - Request with object as a body, do not match content', async (assert) => {
+        const requestOptions = {
+            method: 'POST',
+            body: {
+                0: 1,
+            },
+        };
+        const request = new Request(M3U8_OBJECTS_PATH_01, requestOptions);
+        const MATCH_DATA = 'do_not_match';
+
+        runScriptlet(name, [MATCH_DATA]);
+
+        const done = assert.async();
+
+        const response = await fetch(request);
+        const responseM3U8 = await response.text();
+
+        assert.ok(
+            responseM3U8.includes('tvessaiprod.nbcuni.com/video/'),
+            'content should not been removed',
+        );
+        assert.strictEqual(window.hit, undefined, 'should not hit');
+        done();
     });
 }