+++ tests/scriptlets/xml-prune.test.js
@@ -211,8 +211,10 @@ if (!isSupported) {
                 if (input.includes('trace')) {
                     return;
                 }
-                const EXPECTED_LOG_STR_START = `xml-prune: XMLHttpRequest.open() URL: ${MPD_OBJECTS_PATH}`;
-                assert.ok(input.startsWith(EXPECTED_LOG_STR_START), 'console.hit input');
+                const EXPECTED_LOG_STR_START = 'xml-prune: XMLHttpRequest.open() URL:';
+                const EXPECTED_LOG_PATH = `${(MPD_OBJECTS_PATH).slice(1)}`;
+                assert.ok(input.startsWith(EXPECTED_LOG_STR_START), 'console.hit input EXPECTED_LOG_STR_START');
+                assert.ok(input.includes(EXPECTED_LOG_PATH), 'console.hit input EXPECTED_LOG_PATH');
                 assert.ok(input.includes('pre-roll-1-ad-1'), 'console.hit input');
                 done();
             }
@@ -225,7 +227,6 @@ if (!isSupported) {
         xhr.open(GET_METHOD, MPD_OBJECTS_PATH);
         xhr.onload = () => {
             assert.ok(xhr.responseText.includes('pre-roll-1-ad-1'));
-            assert.strictEqual(window.hit, undefined, 'should not hit');
             done();
         };
         xhr.send();
@@ -245,7 +246,6 @@ if (!isSupported) {
         xhr.open(GET_METHOD, MPD_OBJECTS_PATH);
         xhr.onload = () => {
             assert.ok(xhr.responseText.includes('pre-roll-1-ad-1'));
-            assert.strictEqual(window.hit, undefined, 'should not hit');
             done();
         };
         xhr.send();
@@ -265,7 +265,6 @@ if (!isSupported) {
         xhr.open(GET_METHOD, MPD_OBJECTS_PATH);
         xhr.onload = () => {
             assert.ok(xhr.responseText.includes('pre-roll-1-ad-1'));
-            assert.strictEqual(window.hit, undefined, 'should not hit');
             done();
         };
         xhr.send();
@@ -285,7 +284,6 @@ if (!isSupported) {
         xhr.open(GET_METHOD, MPD_OBJECTS_PATH);
         xhr.onload = () => {
             assert.ok(xhr.responseText.includes('pre-roll-1-ad-1'));
-            assert.strictEqual(window.hit, undefined, 'should not hit');
             done();
         };
         xhr.send();
@@ -363,5 +361,56 @@ if (!isSupported) {
             done();
         };
         xhr.send();
+    });
+
+    test('xhr - remove ads - addEventListener', async (assert) => {
+        const MATCH_DATA = ["Period[id*='-ad-']"];
+
+        runScriptlet(name, MATCH_DATA);
+
+        const done = assert.async();
+
+        const xhr = new XMLHttpRequest();
+        xhr.open(GET_METHOD, MPD_OBJECTS_PATH);
+        xhr.addEventListener('readystatechange', () => {
+            if (xhr.responseText && xhr.readyState >= 3) {
+                assert.notOk(
+                    xhr.responseText.includes('pre-roll-1-ad-1'),
+                    'check if "re-roll-1-ad-1" has been removed',
+                );
+            }
+        });
+        xhr.onload = () => {
+            assert.strictEqual(window.hit, 'FIRED', 'hit function fired');
+            done();
+        };
+        xhr.send();
+    });
+
+    // This test is for issue with - Request with body": Failed to execute 'fetch' on 'Window':
+    // Cannot construct a Request with a Request object that has already been used
+    test('fetch - Request with object as a body', async (assert) => {
+        const requestOptions = {
+            method: 'POST',
+            body: {
+                0: 1,
+            },
+        };
+        const request = new Request(MPD_OBJECTS_PATH, requestOptions);
+        const MATCH_DATA = 'do_not_match';
+
+        runScriptlet(name, [MATCH_DATA]);
+
+        const done = assert.async();
+
+        const response = await fetch(request);
+        const responseMPD = await response.text();
+
+        assert.ok(
+            responseMPD.includes('pre-roll-'),
+            'content should not been removed',
+        );
+        assert.strictEqual(window.hit, undefined, 'should not hit');
+        done();
     });
 }